Dzialajace logowanie do facebooka, dziala pobieranie zdjec profilowych, dziala dodanie maski

+++ SOSWioskiDzieciece.py
@@ -9,28 +9,32 @@ from flask import Flask, request, render_template
 from PIL import Image
 import requests
 import OverlayingImages
-
+#app_url = '/matlaczm/SOS'
 app = Flask(__name__)
 app.debug = True
-token = 'CAACEdEose0cBAJuQXTZB0vlgFCMb4TsW6OWX9Rn71Hm9djZCSclPisYQKk934CQj3vqSjdTdnXNrV8vwM7v8PAYlLVtmYNuMdmToZCYNEVAVdYkZBMRTOnWDs712LhFK466YapSUftEkgPIMIis7LZClHLi5sG3X5OUoY2AmsK4Es8o9p1GZBs79ZAhVAQS0QCUZBPml6i9W1HkzZCvorm5BD'
 
 
 @app.route('/', methods = ['GET', 'POST'])
 def hello_world():
-    qs = 'https://graph.facebook.com/me/picture?type=square&height=3000&width=3000' + '&access_token=' + token
-    res = requests.request('GET', qs)
-    file = cStringIO.StringIO(res.content)
-    img = Image.open(file)
-    img = OverlayingImages.overlay(img)
-    tmp = tempfile.mktemp(suffix='.jpg', dir='./')
-    img.save(tmp)
-    image = open(tmp, 'rb')
-    img = image.read()
-    os.remove('./' + tmp)
-    trueImage = b64encode( img )
+    if request.method == 'POST':
+        token = request.form['token']
+        uid = request.form['uid']
+        qs = 'https://graph.facebook.com/'+ str(uid) +'/picture?type=square&height=3000&width=3000&?access_token=' + str(token)
+        res = requests.request('GET', qs)
+        #print(res.content)
+        file = cStringIO.StringIO(res.content)
+        img = Image.open(file)
+        img = OverlayingImages.overlay(img)
+        img.show()
+        tmp = tempfile.mktemp(suffix='.jpg', dir='./')
+        img.save(tmp)
+        image = open(tmp, 'rb')
+        img = image.read()
+        os.remove('./' + tmp)
+        trueImage = b64encode( img )
 
 
-    return render_template('templates.html', trueImage = trueImage )
+        return render_template('templates.html', trueImage = trueImage )
 
 @app.route('/token', methods=['GET', 'POST'])
 def token():
@@ -41,7 +45,7 @@ def token():
             print "dodajemy post"
         return render_template('operations.html')
     else:
-        return render_template('operations.html')
+        return render_template('facebook.html')
 
 
 if __name__ == '__main__':
